+++ library/ConsoleLib.py
@@ -1,5 +1,6 @@
-import discord
+import sys
 import random
+import discord
 
 from datetime import datetime
 from discord.ext import commands
@@ -151,4 +152,21 @@ class Roundtrip:
     def rt(self):
         roundtrip = round(self.client.latency * 1000)
         return roundtrip
-    
+
+class InitProg:
+    def startProgress(title):
+        global progress_x
+        sys.stdout.write(title + ": [" + "-"*44 + "]" + chr(8)*45)
+        sys.stdout.flush()
+        progress_x = 0
+
+    def progress(x):
+        global progress_x
+        x = int(x * 44 // 100)
+        sys.stdout.write("#" * (x - progress_x))
+        sys.stdout.flush()
+        progress_x = x
+
+    def endProgress():
+        sys.stdout.write("#" * (44 - progress_x) + "]\n")
+        sys.stdout.flush()